+++ test/config.js
@@ -1,8 +1,9 @@
 // @flow
-import { expect } from 'chai';
+import {expect} from 'chai';
 import cookies from 'js-cookie';
+import sinon from 'sinon';
 
-import type { ContactHubFunction } from '../lib/types';
+import type {ContactHubFunction} from '../lib/types';
 
 /* global describe, it, beforeEach */
 
@@ -11,10 +12,9 @@ const varName = 'ch';
 
 const getCookie = () => cookies.getJSON(cookieName) || {};
 
-const _ch:ContactHubFunction = window[varName];
+const _ch: ContactHubFunction = window[varName];
 
 describe('Config API', () => {
-
   beforeEach(() => {
     cookies.remove(cookieName);
   });
@@ -58,6 +58,14 @@ describe('Config API', () => {
     expect(getCookie().context).to.equal('WEB');
   });
 
+  it('sets `{}` if contextInfo not provided', () => {
+    expect(getCookie().contextInfo).to.eql({});
+  });
+
+  it('sets `false` if debug not provided', () => {
+    expect(getCookie().debug).to.equal(false);
+  });
+
   it('allows to specify optional context', () => {
     _ch('config', {
       workspaceId: 'workspace_id',
@@ -74,18 +82,31 @@ describe('Config API', () => {
       nodeId: 'node_id',
       token: 'ABC123',
       context: 'foo',
-      contextInfo: {
-        foo: 'bar'
-      }
+      contextInfo: {foo: 'bar'}
     });
-    expect(getCookie().contextInfo).to.eql({ foo: 'bar' });
+    expect(getCookie().contextInfo).to.eql({foo: 'bar'});
+  });
+
+  it('allows to specify optional debug', () => {
+    _ch('config', {
+      workspaceId: 'workspace_id',
+      nodeId: 'node_id',
+      token: 'ABC123',
+      context: 'foo',
+      contextInfo: {foo: 'bar'},
+      debug: true
+    });
+    expect(getCookie().debug).to.equal(true);
   });
 
   it('removes user data from cookie if the token changes', () => {
-    cookies.set(cookieName, Object.assign(getCookie(), {
-      customerId: 'customer-id',
-      token: 'ABC123'
-    }));
+    cookies.set(
+      cookieName,
+      Object.assign(getCookie(), {
+        customerId: 'customer-id',
+        token: 'ABC123'
+      })
+    );
 
     _ch('config', {
       workspaceId: 'workspace_id',
@@ -96,5 +117,29 @@ describe('Config API', () => {
     expect(getCookie().token).to.equal('CDE456');
     expect(getCookie().customerId).to.be.undefined;
     expect(getCookie().hash).to.be.undefined;
+  });
+
+  it('throws if required option are not specified', () => {
+    expect(() => {
+      // $FlowFixMe
+      _ch('config', {});
+    }).to.throw();
+  });
+
+  it('throws if required option are not specified - debug', () => {
+    const spy = sinon.stub(console, 'error').callsFake(() => undefined);
+
+    expect(() => {
+      // $FlowFixMe
+      _ch('config', {debug: true});
+    }).to.throw();
+
+    expect(
+      spy.calledWith(
+        '[DEBUG] contacthub-sdk-browser',
+        'Invalid ContactHub configuration'
+      )
+    ).to.be.true;
+    spy.restore();
   });
 });