+++ test/customer.js
@@ -1,10 +1,10 @@
 // @flow
 
-import { expect } from 'chai';
+import {expect} from 'chai';
 import cookies from 'js-cookie';
 import sinon from 'sinon';
 
-import type { ContactHubFunction } from '../lib/types';
+import type {ContactHubFunction} from '../lib/types';
 
 /* global describe, it, beforeEach, afterEach */
 
@@ -29,12 +29,16 @@ const mario = {
   }
 };
 
-const _ch:ContactHubFunction = window[varName];
+const _ch: ContactHubFunction = window[varName];
 
 const getCookie = () => cookies.getJSON(cookieName) || {};
 
-const setConfig = () => { _ch('config', config); };
+const setConfig = d => {
+  const debug = d || false;
+  _ch('config', Object.assign({}, config, {debug}));
+};
 
+let spyError;
 let requests = [];
 let xhr;
 
@@ -46,17 +50,22 @@ const whenDone = f => {
   setTimeout(() => f(), 2);
 };
 
+const debugMsg = msg =>
+  spyError.calledWith('[DEBUG] contacthub-sdk-browser', msg);
+
 describe('Customer API:', () => {
   beforeEach(() => {
+    spyError = sinon.stub(console, 'error').callsFake(() => undefined);
     cookies.remove(cookieName);
     requests = [];
     xhr = sinon.useFakeXMLHttpRequest();
-    xhr.onCreate = (xhr) => {
+    xhr.onCreate = xhr => {
       requests.push(xhr);
     };
   });
 
-  afterEach((done) => {
+  afterEach(done => {
+    spyError.restore();
     whenDone(() => done());
   });
 
@@ -75,7 +84,7 @@ describe('Customer API:', () => {
         _ch('customer', mario);
       });
 
-      it('creates a new customer', (done) => {
+      it('creates a new customer', done => {
         whenDone(() => {
           expect(requests.length).to.equal(1);
           const req = requests[0];
@@ -95,21 +104,27 @@ describe('Customer API:', () => {
         });
       });
 
-      it('handles 409 conflicts and updates the existing customer', (done) => {
+      it('handles 409 conflicts and updates the existing customer', done => {
         whenDone(() => {
-          requests[0].respond(409, {}, JSON.stringify({
-            data: {
-              customer: {
-                id: 'existing-cid'
+          requests[0].respond(
+            409,
+            {},
+            JSON.stringify({
+              data: {
+                customer: {
+                  id: 'existing-cid'
+                }
               }
-            }
-          }));
+            })
+          );
           whenDone(() => {
             expect(requests.length).to.equal(2);
             const req = requests[1];
             expect(req.method).to.equal('PATCH');
             expect(req.url).to.equal(
-              `${apiUrl}/workspaces/${config.workspaceId}/customers/existing-cid`
+              `${apiUrl}/workspaces/${
+                config.workspaceId
+              }/customers/existing-cid`
             );
             expect(JSON.parse(req.requestBody)).to.eql({
               externalId: mario.externalId,
@@ -120,9 +135,9 @@ describe('Customer API:', () => {
         });
       });
 
-      it('stores the customerId for future calls', (done) => {
+      it('stores the customerId for future calls', done => {
         whenDone(() => {
-          requests[0].respond(200, {}, JSON.stringify({ id: 'new-cid' }));
+          requests[0].respond(200, {}, JSON.stringify({id: 'new-cid'}));
           whenDone(() => {
             expect(getCookie().customerId).to.equal('new-cid');
             done();
@@ -130,9 +145,9 @@ describe('Customer API:', () => {
         });
       });
 
-      it('stores a hash of the customer data for future calls', (done) => {
+      it('stores a hash of the customer data for future calls', done => {
         whenDone(() => {
-          requests[0].respond(200, {}, JSON.stringify({ id: 'new-cid' }));
+          requests[0].respond(200, {}, JSON.stringify({id: 'new-cid'}));
           whenDone(() => {
             expect(getCookie().hash).not.to.be.undefined;
             done();
@@ -140,15 +155,17 @@ describe('Customer API:', () => {
         });
       });
 
-      it('reconciles the sessionId with the customerId', (done) => {
+      it('reconciles the sessionId with the customerId', done => {
         const sid = getCookie().sid;
         whenDone(() => {
-          requests[0].respond(200, {}, JSON.stringify({ id: 'new-cid' }));
+          requests[0].respond(200, {}, JSON.stringify({id: 'new-cid'}));
           whenDone(() => {
             expect(requests.length).to.equal(2);
             const req = requests[1];
             expect(req.url).to.equal(
-              `${apiUrl}/workspaces/${config.workspaceId}/customers/new-cid/sessions`
+              `${apiUrl}/workspaces/${
+                config.workspaceId
+              }/customers/new-cid/sessions`
             );
             expect(JSON.parse(req.requestBody)).to.eql({
               value: sid
@@ -162,9 +179,12 @@ describe('Customer API:', () => {
     describe('and a customerId is stored in the cookie', () => {
       beforeEach(() => {
         setConfig();
-        cookies.set(cookieName, Object.assign(getCookie(), {
-          customerId: 'my-cid'
-        }));
+        cookies.set(
+          cookieName,
+          Object.assign(getCookie(), {
+            customerId: 'my-cid'
+          })
+        );
         _ch('customer', mario);
       });
 
@@ -181,7 +201,7 @@ describe('Customer API:', () => {
         });
       });
 
-      it('does not update the customer if the same data is sent', (done) => {
+      it('does not update the customer if the same data is sent', done => {
         requests[0].respond(200);
         whenDone(() => {
           _ch('customer', mario);
@@ -190,7 +210,7 @@ describe('Customer API:', () => {
         });
       });
 
-      it('does update the customer if updated data is sent', (done) => {
+      it('does update the customer if updated data is sent', done => {
         requests[0].respond(200);
         whenDone(() => {
           if (mario.base) {
@@ -202,7 +222,7 @@ describe('Customer API:', () => {
         });
       });
 
-      it('does update the customer if only the externalId has changed', (done) => {
+      it('does update the customer if only the externalId has changed', done => {
         requests[0].respond(200);
         whenDone(() => {
           mario.externalId = 'supermario';
@@ -218,10 +238,13 @@ describe('Customer API:', () => {
     describe('and the same customerId is stored in the cookie', () => {
       beforeEach(() => {
         setConfig();
-        cookies.set(cookieName, Object.assign(getCookie(), {
-          customerId: 'my-cid'
-        }));
-        _ch('customer', { id: 'my-cid' });
+        cookies.set(
+          cookieName,
+          Object.assign(getCookie(), {
+            customerId: 'my-cid'
+          })
+        );
+        _ch('customer', {id: 'my-cid'});
       });
 
       it('does not make any API call', () => {
@@ -232,10 +255,13 @@ describe('Customer API:', () => {
     describe('and a different customerId is stored in the cookie', () => {
       beforeEach(() => {
         setConfig();
-        cookies.set(cookieName, Object.assign(getCookie(), {
-          customerId: 'different-cid'
-        }));
-        _ch('customer', { id: 'my-cid' });
+        cookies.set(
+          cookieName,
+          Object.assign(getCookie(), {
+            customerId: 'different-cid'
+          })
+        );
+        _ch('customer', {id: 'my-cid'});
       });
 
       it('does not make any API call', () => {
@@ -249,21 +275,23 @@ describe('Customer API:', () => {
       });
 
       it('does not reset the sessionId', () => {
-        const { sid } = getCookie();
-        _ch('customer', { id: 'my-cid' });
+        const {sid} = getCookie();
+        _ch('customer', {id: 'my-cid'});
 
         expect(getCookie().sid).to.eql(sid);
       });
 
-      it('reconciles the sessionId with the customerId', (done) => {
-        const { sid } = getCookie();
-        _ch('customer', { id: 'my-cid' });
+      it('reconciles the sessionId with the customerId', done => {
+        const {sid} = getCookie();
+        _ch('customer', {id: 'my-cid'});
 
         whenDone(() => {
           expect(requests.length).to.equal(1);
           const req = requests[0];
           expect(req.url).to.equal(
-            `${apiUrl}/workspaces/${config.workspaceId}/customers/my-cid/sessions`
+            `${apiUrl}/workspaces/${
+              config.workspaceId
+            }/customers/my-cid/sessions`
           );
           expect(JSON.parse(req.requestBody)).to.eql({
             value: sid
@@ -273,7 +301,7 @@ describe('Customer API:', () => {
       });
 
       it('does not make additional API requests', () => {
-        _ch('customer', { id: 'my-cid' });
+        _ch('customer', {id: 'my-cid'});
 
         expect(requests.length).to.equal(1);
       });
@@ -284,13 +312,16 @@ describe('Customer API:', () => {
     describe('and the same customerId is stored in the cookie', () => {
       beforeEach(() => {
         setConfig();
-        cookies.set(cookieName, Object.assign(getCookie(), {
-          customerId: 'my-cid'
-        }));
-        _ch('customer', Object.assign({}, mario, { id: 'my-cid' }));
+        cookies.set(
+          cookieName,
+          Object.assign(getCookie(), {
+            customerId: 'my-cid'
+          })
+        );
+        _ch('customer', Object.assign({}, mario, {id: 'my-cid'}));
       });
 
-      it('updates the customer', (done) => {
+      it('updates the customer', done => {
         whenDone(() => {
           expect(requests.length).to.equal(1);
           const req = requests[0];
@@ -310,27 +341,32 @@ describe('Customer API:', () => {
     describe('and a different customerId is stored in the cookie', () => {
       beforeEach(() => {
         setConfig();
-        cookies.set(cookieName, Object.assign(getCookie(), {
-          customerId: 'different-cid'
-        }));
+        cookies.set(
+          cookieName,
+          Object.assign(getCookie(), {
+            customerId: 'different-cid'
+          })
+        );
       });
 
       it('resets the sessionId', () => {
-        const { sid } = getCookie();
-        _ch('customer', Object.assign({}, mario, { id: 'my-cid' }));
+        const {sid} = getCookie();
+        _ch('customer', Object.assign({}, mario, {id: 'my-cid'}));
 
         expect(getCookie().sid).not.to.eql(sid);
       });
 
-      it('reconciles the new sessionId with the customerId', (done) => {
-        _ch('customer', Object.assign({}, mario, { id: 'my-cid' }));
-        const { sid: newSid } = getCookie();
+      it('reconciles the new sessionId with the customerId', done => {
+        _ch('customer', Object.assign({}, mario, {id: 'my-cid'}));
+        const {sid: newSid} = getCookie();
 
         whenDone(() => {
           expect(requests.length).to.equal(1);
           const req = requests[0];
           expect(req.url).to.equal(
-            `${apiUrl}/workspaces/${config.workspaceId}/customers/my-cid/sessions`
+            `${apiUrl}/workspaces/${
+              config.workspaceId
+            }/customers/my-cid/sessions`
           );
           expect(JSON.parse(req.requestBody)).to.eql({
             value: newSid
@@ -339,8 +375,8 @@ describe('Customer API:', () => {
         });
       });
 
-      it('updates the customer', (done) => {
-        _ch('customer', Object.assign({}, mario, { id: 'my-cid' }));
+      it('updates the customer', done => {
+        _ch('customer', Object.assign({}, mario, {id: 'my-cid'}));
 
         requests[0].respond(200);
         whenDone(() => {
@@ -365,21 +401,23 @@ describe('Customer API:', () => {
       });
 
       it('does not reset the sessionId', () => {
-        const { sid } = getCookie();
-        _ch('customer', Object.assign({}, mario, { id: 'my-cid' }));
+        const {sid} = getCookie();
+        _ch('customer', Object.assign({}, mario, {id: 'my-cid'}));
 
         expect(getCookie().sid).to.eql(sid);
       });
 
-      it('reconciles the sessionId with the customerId', (done) => {
-        const { sid } = getCookie();
-        _ch('customer', Object.assign({}, mario, { id: 'my-cid' }));
+      it('reconciles the sessionId with the customerId', done => {
+        const {sid} = getCookie();
+        _ch('customer', Object.assign({}, mario, {id: 'my-cid'}));
 
         whenDone(() => {
           expect(requests.length).to.equal(1);
           const req = requests[0];
           expect(req.url).to.equal(
-            `${apiUrl}/workspaces/${config.workspaceId}/customers/my-cid/sessions`
+            `${apiUrl}/workspaces/${
+              config.workspaceId
+            }/customers/my-cid/sessions`
           );
           expect(JSON.parse(req.requestBody)).to.eql({
             value: sid
@@ -388,8 +426,8 @@ describe('Customer API:', () => {
         });
       });
 
-      it('updates the customer', (done) => {
-        _ch('customer', Object.assign({}, mario, { id: 'my-cid' }));
+      it('updates the customer', done => {
+        _ch('customer', Object.assign({}, mario, {id: 'my-cid'}));
 
         requests[0].respond(200);
         whenDone(() => {
@@ -413,10 +451,13 @@ describe('Customer API:', () => {
     beforeEach(() => {
       setConfig();
 
-      cookies.set(cookieName, Object.assign(getCookie(), {
-        customerId: 'old-customer-id',
-        sid: 'old-session-id'
-      }));
+      cookies.set(
+        cookieName,
+        Object.assign(getCookie(), {
+          customerId: 'old-customer-id',
+          sid: 'old-session-id'
+        })
+      );
 
       _ch('customer');
     });
@@ -429,6 +470,156 @@ describe('Customer API:', () => {
       expect(getCookie().sid).to.match(
         /^[0-9A-F]{8}-[0-9A-F]{4}-[4][0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$/i
       );
+    });
+  });
+
+  // --- Rejections
+  describe('when id and customerId are provided it should logs catched promise rejection', () => {
+    beforeEach(() => {
+      setConfig(true);
+      cookies.set(
+        cookieName,
+        Object.assign(getCookie(), {
+          customerId: 'TEST'
+        })
+      );
+    });
+
+    it('when id != customerId and no other customer data provided', done => {
+      _ch('customer', {id: 'NO_TEST'});
+
+      whenDone(() => {
+        expect(
+          debugMsg('The provided id conflicts with the id stored in the cookie')
+        ).to.be.true;
+        done();
+      });
+    });
+
+    it('when sessions api respond with error', done => {
+      _ch('customer', {id: 'NO_TEST', externalId: 'ANOTHER_ID'});
+
+      requests[0].respond(500, {}, 'KO');
+
+      whenDone(() => {
+        expect(debugMsg('KO')).to.be.true;
+        done();
+      });
+    });
+
+    it('when customers api respond with error', done => {
+      _ch('customer', {id: 'TEST', externalId: 'ANOTHER_ID'});
+
+      whenDone(() => {
+        requests[0].respond(500, {}, 'KO');
+
+        whenDone(() => {
+          expect(debugMsg('KO')).to.be.true;
+          done();
+        });
+      });
+    });
+  });
+
+  describe('when customerId is NOT provided it should logs catched promise rejection', () => {
+    beforeEach(() => {
+      setConfig(true);
+    });
+
+    it('when sessions api respond with error', done => {
+      _ch('customer', {id: 'TEST'});
+
+      requests[0].respond(500, {}, 'KO');
+
+      whenDone(() => {
+        expect(debugMsg('KO')).to.be.true;
+        done();
+      });
+    });
+
+    it('when customers api respond with error', done => {
+      _ch('customer', {id: 'TEST'});
+
+      whenDone(() => {
+        requests[0].respond(500, {}, 'KO');
+
+        whenDone(() => {
+          expect(debugMsg('KO')).to.be.true;
+          done();
+        });
+      });
+    });
+  });
+
+  describe('when customerId is provided and id is not provided it should logs catched promise rejection', () => {
+    beforeEach(() => {
+      setConfig(true);
+      cookies.set(
+        cookieName,
+        Object.assign(getCookie(), {
+          customerId: 'TEST'
+        })
+      );
+    });
+
+    it('when customers api respond with error', done => {
+      _ch('customer', {externalId: 'ANOTHER_ID'});
+
+      requests[0].respond(500, {}, 'KO');
+
+      whenDone(() => {
+        whenDone(() => {
+          expect(debugMsg('KO')).to.be.true;
+          done();
+        });
+      });
+    });
+  });
+
+  describe('when customerId and id are not provided it should logs catched promise rejection', () => {
+    beforeEach(() => {
+      setConfig(true);
+    });
+
+    it('when customers api respond with error', done => {
+      _ch('customer', {});
+
+      requests[0].respond(500, {}, 'KO');
+
+      whenDone(() => {
+        whenDone(() => {
+          expect(debugMsg('KO')).to.be.true;
+          done();
+        });
+      });
+    });
+
+    it('when customer api respond with error (on merge)', done => {
+      _ch('customer', {});
+
+      whenDone(() => {
+        requests[0].respond(500, {}, 'KO');
+
+        whenDone(() => {
+          expect(debugMsg('KO')).to.be.true;
+          done();
+        });
+      });
+    });
+
+    it('when sessions api respond with error', done => {
+      _ch('customer', {});
+
+      whenDone(() => {
+        whenDone(() => {
+          requests[0].respond(500, {}, 'KO');
+
+          whenDone(() => {
+            expect(debugMsg('KO')).to.be.true;
+            done();
+          });
+        });
+      });
     });
   });
 });