+++ CHANGELOG.md
@@ -1,8 +1,17 @@
-#  Change Log
+# Change Log
 
+## [v1.4.0](https://github.com/contactlab/contacthub-sdk-browser/tree/v1.4.0) (2019-06-07)
 
+[Full Changelog](https://github.com/contactlab/contacthub-sdk-browser/compare/v1.3.0...v1.4.0)
+
+This version adds a `debug` flag in configuration object which if it is set to `true` will lead the SDK to log errors to the browser's console.
+
+#### Fixes (bugs & defects):
+
+- Calls to Hub api via xr do not catch promises rejections [#44](https://github.com/contactlab/contacthub-sdk-browser/issues/44)
 
 ## [v1.3.0](https://github.com/contactlab/contacthub-sdk-browser/tree/v1.3.0) (2018-05-25)
+
 [Full Changelog](https://github.com/contactlab/contacthub-sdk-browser/compare/v1.2.0...v1.3.0)
 
 #### New features:
@@ -11,6 +20,7 @@
 - Expire utm information after 30 minutes [#39](https://github.com/contactlab/contacthub-sdk-browser/issues/39)
 
 ## [v1.2.0](https://github.com/contactlab/contacthub-sdk-browser/tree/v1.2.0) (2018-04-30)
+
 [Full Changelog](https://github.com/contactlab/contacthub-sdk-browser/compare/v1.1.1...v1.2.0)
 
 #### New features:
@@ -18,6 +28,7 @@
 - Support new "consents" object for Customers [#36](https://github.com/contactlab/contacthub-sdk-browser/issues/36)
 
 ## [v1.1.1](https://github.com/contactlab/contacthub-sdk-browser/tree/v1.1.1) (2017-08-30)
+
 [Full Changelog](https://github.com/contactlab/contacthub-sdk-browser/compare/v1.1.0...v1.1.1)
 
 #### Fixes (bugs & defects):
@@ -25,6 +36,7 @@
 - window.Promise should not be modified [#34](https://github.com/contactlab/contacthub-sdk-browser/issues/34)
 
 ## [v1.1.0](https://github.com/contactlab/contacthub-sdk-browser/tree/v1.1.0) (2017-06-01)
+
 [Full Changelog](https://github.com/contactlab/contacthub-sdk-browser/compare/v1.0.0...v1.1.0)
 
 #### Fixes (bugs & defects):
@@ -36,6 +48,7 @@
 - Allow using an existing customerId [#30](https://github.com/contactlab/contacthub-sdk-browser/issues/30)
 
 ## [v1.0.0](https://github.com/contactlab/contacthub-sdk-browser/tree/v1.0.0) (2017-03-17)
+
 [Full Changelog](https://github.com/contactlab/contacthub-sdk-browser/compare/v0.4.1...v1.0.0)
 
 #### New features:
@@ -44,6 +57,7 @@
 - run build on Travis [#24](https://github.com/contactlab/contacthub-sdk-browser/issues/24)
 
 ## [v0.4.1](https://github.com/contactlab/contacthub-sdk-browser/tree/v0.4.1) (2016-12-19)
+
 [Full Changelog](https://github.com/contactlab/contacthub-sdk-browser/compare/v0.4.0...v0.4.1)
 
 #### Fixes (bugs & defects):
@@ -51,6 +65,7 @@
 - Update to work with new API version (without hyperlinks) [#22](https://github.com/contactlab/contacthub-sdk-browser/issues/22)
 
 ## [v0.4.0](https://github.com/contactlab/contacthub-sdk-browser/tree/v0.4.0) (2016-11-17)
+
 [Full Changelog](https://github.com/contactlab/contacthub-sdk-browser/compare/v0.3.0...v0.4.0)
 
 #### Fixes (bugs & defects):
@@ -65,6 +80,7 @@
 - Handle 409 API errors on create [#12](https://github.com/contactlab/contacthub-sdk-browser/issues/12)
 
 ## [v0.3.0](https://github.com/contactlab/contacthub-sdk-browser/tree/v0.3.0) (2016-11-02)
+
 [Full Changelog](https://github.com/contactlab/contacthub-sdk-browser/compare/v0.2.2...v0.3.0)
 
 ## [v0.2.2](https://github.com/contactlab/contacthub-sdk-browser/tree/v0.2.2) (2016-10-07)